+++ tests/did/01-did-create.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { sign } from '@stablelib/ed25519';
-import { toString, fromString } from 'uint8arrays';
+import { toString } from 'uint8arrays';
 import { getDidDocument, privKeyBytes, pubKeyHex, setDidDocument } from 'fixtures';
 
 let didState;
@@ -49,17 +49,19 @@ test('did-create. Initiate DID Create procedure', async ({ request }) => {
 
 test('did-create. Send the final request for DID creating', async ({ request }) => {
 	const didPayload = getDidDocument();
-	const serializedPayload = didState.signingRequest[0].serializedPayload;
+	const signingRequest = didState.signingRequest['signingRequest0'];
+
+	const serializedPayload = signingRequest.serializedPayload;
 	const serializedBytes = Buffer.from(serializedPayload, 'base64');
 	const signature = sign(privKeyBytes, serializedBytes);
 
 	const secret = {
-		signingResponse: [
-			{
-				kid: didState.signingRequest[0].kid,
+		signingResponse: {
+			signingRequest0: {
+				kid: signingRequest.kid,
 				signature: toString(signature, 'base64'),
 			},
-		],
+		},
 	};
 
 	const didCreate = await request.post(`/1.0/create/`, {