+++ tests/did/03-did-deactivate.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { sign } from '@stablelib/ed25519';
-import { toString, fromString } from 'uint8arrays';
+import { toString } from 'uint8arrays';
 import { getDidDocument, privKeyBytes } from 'fixtures';
 
 let didState;
@@ -34,17 +34,18 @@ test('did-deactivate. Initiate DID Deactivate procedure', async ({ request }) =>
 
 test('did-deactivate. Send the final request for DID deactivating', async ({ request }) => {
 	const didPayload = getDidDocument();
-	const serializedPayload = didState.signingRequest[0].serializedPayload;
+	const signingRequest = didState.signingRequest['signingRequest0'];
+	const serializedPayload = signingRequest.serializedPayload;
 	const serializedBytes = Buffer.from(serializedPayload, 'base64');
 	const signature = sign(privKeyBytes, serializedBytes);
 
 	const secret = {
-		signingResponse: [
-			{
-				kid: didState.signingRequest[0].kid,
+		signingResponse: {
+			signingRequest0: {
+				kid: signingRequest.kid,
 				signature: toString(signature, 'base64'),
 			},
-		],
+		},
 	};
 
 	const didDeactivate = await request.post(`/1.0/deactivate`, {