+++ tests/resource/02-createResource.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { sign } from '@stablelib/ed25519';
-import { toString, fromString } from 'uint8arrays';
+import { toString } from 'uint8arrays';
 import { getDidDocument, privKeyBytes, setResourceId } from 'fixtures';
 
 let didUrlState;
@@ -12,11 +12,11 @@ test('resource-create. Initiate Resource creation procedure', async ({ request }
 		data: {
 			did: didPayload.id,
 			content: 'SGVsbG8gV29ybGQ=',
-			name: 'ResourceName',
-			type: 'TextDocument',
-			version: '1.0',
 			options: {
 				network: 'testnet',
+				name: 'ResourceName',
+				type: 'TextDocument',
+				versionId: '1.0',
 			},
 		},
 	});
@@ -35,30 +35,31 @@ test('resource-create. Initiate Resource creation procedure', async ({ request }
 
 test('resource-create. Send the final request for Resource creation', async ({ request }) => {
 	const didPayload = getDidDocument();
-	const serializedPayload = didUrlState.signingRequest[0].serializedPayload;
+	const signingRequest = didUrlState.signingRequest['signingRequest0'];
+	const serializedPayload = signingRequest.serializedPayload;
 	const serializedBytes = Buffer.from(serializedPayload, 'base64');
 	const signature = sign(privKeyBytes, serializedBytes);
 
 	const secret = {
-		signingResponse: [
-			{
-				kid: didUrlState.signingRequest[0].kid,
+		signingResponse: {
+			signingRequest0: {
+				kid: signingRequest.kid,
 				signature: toString(signature, 'base64'),
 			},
-		],
+		},
 	};
 
 	const resourceCreate = await request.post(`/1.0/createResource`, {
 		data: {
 			did: didPayload.id,
 			content: 'SGVsbG8gV29ybGQ=',
-			name: 'ResourceName',
-			type: 'TextDocument',
-			version: '1.0',
 			jobId: resourceJobId,
 			secret: secret,
 			options: {
 				network: 'testnet',
+				name: 'ResourceName',
+				type: 'TextDocument',
+				versionId: '1.0',
 			},
 		},
 	});