+++ tests/resource/03-updateResource.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { sign } from '@stablelib/ed25519';
-import { toString, fromString } from 'uint8arrays';
+import { toString } from 'uint8arrays';
 import { getDidDocument, getResourceId, privKeyBytes } from 'fixtures';
 
 let didUrlState;
@@ -12,13 +12,13 @@ test('resource-update. Initiate Resource update procedure', async ({ request }) 
 	const payload = await request.post(`/1.0/updateResource`, {
 		data: {
 			did: didPayload.id,
-			name: 'ResourceName',
-			type: 'TextDocument',
 			content: ['SGVsbG8gV29ybGQ='],
-			version: '2.0',
 			relativeDidUrl: '/resources/' + resourceId,
 			options: {
 				network: 'testnet',
+				name: 'ResourceName',
+				type: 'TextDocument',
+				versionId: '2.0',
 			},
 		},
 	});
@@ -38,31 +38,32 @@ test('resource-update. Initiate Resource update procedure', async ({ request }) 
 test('resource-update. Send the final request for Resource update', async ({ request }) => {
 	const didPayload = getDidDocument();
 	const resourceId = getResourceId();
-	const serializedPayload = didUrlState.signingRequest[0].serializedPayload;
+	const signingRequest = didUrlState.signingRequest['signingRequest0'];
+	const serializedPayload = signingRequest.serializedPayload;
 	const serializedBytes = Buffer.from(serializedPayload, 'base64');
 	const signature = sign(privKeyBytes, serializedBytes);
 
 	const secret = {
-		signingResponse: [
-			{
-				kid: didUrlState.signingRequest[0].kid,
+		signingResponse: {
+			signingRequest0: {
+				kid: signingRequest.kid,
 				signature: toString(signature, 'base64'),
 			},
-		],
+		},
 	};
 
 	const resourceUpdate = await request.post(`/1.0/updateResource`, {
 		data: {
 			did: didPayload.id,
 			content: ['SGVsbG8gV29ybGQ='],
-			name: 'ResourceName',
-			type: 'TextDocument',
-			version: '1.0',
 			jobId: resourceJobId,
 			relativeDidUrl: '/resources/' + resourceId,
 			secret: secret,
 			options: {
 				network: 'testnet',
+				name: 'ResourceName',
+				type: 'TextDocument',
+				versionId: '1.0',
 			},
 		},
 	});
@@ -82,12 +83,12 @@ test('resource-update. Resource update without relativeDidUrl', async ({ request
 	const payload = await request.post(`/1.0/updateResource`, {
 		data: {
 			did: didPayload.id,
-			name: 'ResourceName',
-			type: 'TextDocument',
 			content: ['SGVsbG8gV29ybGQ='],
-			version: '3.0',
 			options: {
 				network: 'testnet',
+				name: 'ResourceName',
+				type: 'TextDocument',
+				versionId: '3.0',
 			},
 		},
 	});
@@ -106,17 +107,18 @@ test('resource-update. Resource update without relativeDidUrl', async ({ request
 
 test('resource-update. Send the final update without relativeDidUrl', async ({ request }) => {
 	const didPayload = getDidDocument();
-	const serializedPayload = didUrlState.signingRequest[0].serializedPayload;
+	const signingRequest = didUrlState.signingRequest['signingRequest0'];
+	const serializedPayload = signingRequest.serializedPayload;
 	const serializedBytes = Buffer.from(serializedPayload, 'base64');
 	const signature = sign(privKeyBytes, serializedBytes);
 
 	const secret = {
-		signingResponse: [
-			{
-				kid: didUrlState.signingRequest[0].kid,
+		signingResponse: {
+			signingRequest0: {
+				kid: signingRequest.kid,
 				signature: toString(signature, 'base64'),
 			},
-		],
+		},
 	};
 
 	const resourceUpdate = await request.post(`/1.0/updateResource`, {